db: triggers: split the trigger repository
Split the trigger repository based on permissions:

	- ITriggerAdministrationRepository for the Network API to manage
	  triggers;
	- ITriggerRepository for the trigger service to execute triggers.
